# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorEnvelope {
  error: string;
}

export type ShowSourceType = typeof ShowSourceType[keyof typeof ShowSourceType];


export const ShowSourceType = {
  uploaded: 'uploaded',
  youtube: 'youtube',
} as const;

export type MediaType = typeof MediaType[keyof typeof MediaType];


export const MediaType = {
  movie: 'movie',
  series: 'series',
} as const;

export interface Show {
  id: number;
  title: string;
  slug: string;
  synopsis: string;
  genres: string[];
  year: number;
  rating: number;
  episodesCount: number;
  mediaType: MediaType;
  /** @nullable */
  thumbnailUrl: string | null;
  /** @nullable */
  bannerUrl: string | null;
  sourceType: ShowSourceType;
  /** @nullable */
  videoUrl: string | null;
  /** @nullable */
  videoPath: string | null;
  featured: boolean;
  createdAt: string;
}

export interface ShowInput {
  /** @minLength 1 */
  title: string;
  synopsis: string;
  genres: string[];
  year: number;
  rating: number;
  episodesCount: number;
  mediaType: MediaType;
  /** @nullable */
  thumbnailUrl?: string | null;
  /** @nullable */
  bannerUrl?: string | null;
  sourceType: ShowSourceType;
  /** @nullable */
  videoUrl?: string | null;
  /** @nullable */
  videoPath?: string | null;
  featured: boolean;
}

export interface ShowUpdate {
  /** @minLength 1 */
  title?: string;
  synopsis?: string;
  genres?: string[];
  year?: number;
  rating?: number;
  episodesCount?: number;
  mediaType?: MediaType;
  /** @nullable */
  thumbnailUrl?: string | null;
  /** @nullable */
  bannerUrl?: string | null;
  sourceType?: ShowSourceType;
  /** @nullable */
  videoUrl?: string | null;
  /** @nullable */
  videoPath?: string | null;
  featured?: boolean;
}

export interface Highlights {
  featured: Show[];
  trending: Show[];
  latest: Show[];
}

export interface Season {
  id: number;
  showId: number;
  seasonNumber: number;
  title: string;
  createdAt: string;
}

export interface SeasonInput {
  /** @minimum 1 */
  seasonNumber: number;
  title?: string;
}

export interface Episode {
  id: number;
  seasonId: number;
  episodeNumber: number;
  title: string;
  synopsis: string;
  sourceType: ShowSourceType;
  /** @nullable */
  videoUrl: string | null;
  /** @nullable */
  videoPath: string | null;
  createdAt: string;
}

export interface EpisodeInput {
  /** @minimum 1 */
  episodeNumber: number;
  /** @minLength 1 */
  title: string;
  synopsis: string;
  sourceType: ShowSourceType;
  /** @nullable */
  videoUrl?: string | null;
  /** @nullable */
  videoPath?: string | null;
}

export interface DeveloperLock {
  enabled: boolean;
  configured: boolean;
}

export interface DeveloperLockInput {
  enabled: boolean;
  /**
     * @minLength 4
     * @nullable
     */
  password?: string | null;
}

export interface DeveloperLockVerify {
  /** @minLength 1 */
  password: string;
}

export interface DeveloperLockVerification {
  verified: boolean;
}

export interface UploadUrlRequest {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
  metadata: UploadUrlRequest;
}

export type ListShowsParams = {
query?: string;
genre?: string;
sourceType?: ShowSourceType;
};