# lib/api-zod/src/generated/types/episode.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ShowSourceType } from './showSourceType';

export interface Episode {
  id: number;
  seasonId: number;
  episodeNumber: number;
  title: string;
  synopsis: string;
  sourceType: ShowSourceType;
  /** @nullable */
  videoUrl: string | null;
  /** @nullable */
  videoPath: string | null;
  createdAt: Date;
}