# lib/api-zod/src/generated/types/mediaType.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type MediaType = typeof MediaType[keyof typeof MediaType];


export const MediaType = {
  movie: 'movie',
  series: 'series',
} as const;