# lib/api-zod/src/generated/types/season.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Season {
  id: number;
  showId: number;
  seasonNumber: number;
  title: string;
  createdAt: Date;
}