# lib/api-zod/src/generated/types/seasonInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SeasonInput {
  /** @minimum 1 */
  seasonNumber: number;
  title?: string;
}